# Code optimizations for speed