Thanks for giving me access @jeanpauphilet! I was reading through the code to get familiar with the algorithm before we think about how to use it in Optimal Trees and noticed some small things in the code that were slowing it down. This PR has changes to fix those things that were happening in the hot path of the algorithm and lead to some pretty significant speedups (~3-4x) and reduction in memory usage (~1000x). Most of the speedups work on both julia 0.5 and 0.6, whereas the memory reduction relies on the improvements to [broadcasting in 0.6](https://julialang.org/blog/2017/01/moredots). I've included speed and memory comparisons below. I only considered OLS for now.

Test code:
```julia
using SubsetSelection
using StatsBase
using Base.Test

srand(10)
n = 5000; p = 10000; k = 10;

indices = sort(sample(1:p, StatsBase.Weights(ones(p)/p), k, replace=false));
w = sample(-1:2:1, k);
X = randn(n,p); Y = X[:,indices]*w;

@time Sparse_Regressor = SubsetSelection.subsetSelection(SubsetSelection.OLS(), SubsetSelection.Constraint(10), Y, X, intercept=true)
@time Sparse_Regressor = SubsetSelection.subsetSelection(SubsetSelection.OLS(), SubsetSelection.Constraint(10), Y, X, intercept=false)
@time Sparse_Regressor = SubsetSelection.subsetSelection(SubsetSelection.OLS(), SubsetSelection.Penalty(10000.0), Y, X, intercept=true)
@time Sparse_Regressor = SubsetSelection.subsetSelection(SubsetSelection.OLS(), SubsetSelection.Penalty(10000.0), Y, X, intercept=false)
```

Before - Julia 0.5:
```
  7.601780 seconds (196.27 k allocations: 1.764 GB, 4.10% gc time)
  7.371120 seconds (148.32 k allocations: 1.616 GB, 2.31% gc time)
 13.626151 seconds (2.99 M allocations: 38.130 GB, 21.63% gc time)
 14.106672 seconds (3.03 M allocations: 38.700 GB, 22.31% gc time)
```

After - Julia 0.5:
```
  2.939549 seconds (741.97 k allocations: 879.339 MB, 8.90% gc time)
  2.753066 seconds (739.25 k allocations: 841.021 MB, 3.29% gc time)
  2.972226 seconds (1.11 M allocations: 1.237 GB, 4.59% gc time)
  3.181590 seconds (1.47 M allocations: 1.567 GB, 5.07% gc time)
```

After - Julia 0.6:

```
  2.143580 seconds (10.56 k allocations: 1.339 MiB)
  2.263023 seconds (10.54 k allocations: 1.262 MiB)
  2.153016 seconds (14.94 k allocations: 1.773 MiB)
  2.141344 seconds (19.73 k allocations: 2.065 MiB)
```

Let me know what you think and feel free to drop by to discuss!